clean up extra code
