+++ components/charts/data/__tests__/getBarData.test.ts
@@ -1,11 +1,12 @@
 import { faker } from "@faker-js/faker";
 import { describe, expect, test } from "vitest";
-import { _getBarData, type ITabData } from "../getBarData";
+import { mockTab } from "@/test-utils/mockDataHelper";
+import { _getBarData } from "../getBarData";
 
 describe("getBarData", () => {
   describe("_addChildren", () => {
     test("should map child data for chart with depth 0", () => {
-      const tabs: Array<ITabData> = [];
+      const tabs: Array<Browser.tabs.Tab> = [];
 
       const origin1 = faker.internet.url({ appendSlash: false });
       const origin2 = faker.internet.url({ appendSlash: false });
@@ -14,9 +15,9 @@ describe("getBarData", () => {
       const segment2 = faker.internet.domainWord();
       const segment3 = faker.internet.domainWord();
 
-      tabs.push({ url: `${origin1}/${segment1}` });
-      tabs.push({ url: `${origin1}/${segment2}` });
-      tabs.push({ url: `${origin2}/${segment3}` });
+      tabs.push(mockTab({ url: `${origin1}/${segment1}` }));
+      tabs.push(mockTab({ url: `${origin1}/${segment2}` }));
+      tabs.push(mockTab({ url: `${origin2}/${segment3}` }));
 
       const res = _getBarData(tabs, 0, 10, 0);
 
@@ -26,7 +27,7 @@ describe("getBarData", () => {
     });
 
     test("should map child data for chart with depth 1", () => {
-      const tabs: Array<ITabData> = [];
+      const tabs: Array<Browser.tabs.Tab> = [];
 
       const origin1 = faker.internet.url({ appendSlash: false });
       const origin2 = faker.internet.url({ appendSlash: false });
@@ -39,10 +40,10 @@ describe("getBarData", () => {
       const base2 = `${origin1}/${segment2}`;
       const base3 = `${origin2}/${segment3}`;
 
-      tabs.push({ url: `${base1}/0` });
-      tabs.push({ url: `${base2}/0` });
-      tabs.push({ url: `${base3}/1` });
-      tabs.push({ url: `${base3}/2` });
+      tabs.push(mockTab({ url: `${base1}/0` }));
+      tabs.push(mockTab({ url: `${base2}/0` }));
+      tabs.push(mockTab({ url: `${base3}/1` }));
+      tabs.push(mockTab({ url: `${base3}/2` }));
 
       const res = _getBarData(tabs, 1, 10, 0);
 
@@ -53,7 +54,7 @@ describe("getBarData", () => {
     });
 
     test("should map child data for chart with depth with depth 2", () => {
-      const tabs: Array<ITabData> = [];
+      const tabs: Array<Browser.tabs.Tab> = [];
 
       const origin1 = faker.internet.url({ appendSlash: false });
       const origin2 = faker.internet.url({ appendSlash: false });
@@ -67,11 +68,11 @@ describe("getBarData", () => {
       const base3 = `${origin2}/${segment2}/${segment2}`;
       const base4 = `${origin2}/${segment2}`;
 
-      tabs.push({ url: `${base1}/0` });
-      tabs.push({ url: `${base2}/0` });
-      tabs.push({ url: `${base3}/1` });
-      tabs.push({ url: `${base3}/2` });
-      tabs.push({ url: `${base4}` });
+      tabs.push(mockTab({ url: `${base1}/0` }));
+      tabs.push(mockTab({ url: `${base2}/0` }));
+      tabs.push(mockTab({ url: `${base3}/1` }));
+      tabs.push(mockTab({ url: `${base3}/2` }));
+      tabs.push(mockTab({ url: `${base4}` }));
 
       const res = _getBarData(tabs, 2, 10, 0);
