--- components/tabs/types.ts
@@ -1,5 +1,0 @@
-export enum TabValue {
-  Tab = "tab",
-  Count = "count",
-  Snooze = "snooze",
-}
