--- components/ui/select.tsx
@@ -1,35 +1,0 @@
-import { forwardRef, type SelectHTMLAttributes } from "react";
-import { cn } from "@/lib/utils";
-
-export interface SelectOption {
-  label: string;
-  value: string;
-}
-
-export interface SelectProps extends SelectHTMLAttributes<HTMLSelectElement> {
-  options: SelectOption[];
-}
-
-const Select = forwardRef<HTMLSelectElement, SelectProps>(function Select(
-  { className, options, ...props },
-  ref,
-) {
-  return (
-    <select
-      ref={ref}
-      className={cn(
-        "h-9 min-w-40 rounded-md border border-input bg-background px-3 text-sm outline-none transition-colors focus:border-ring focus:ring-2 focus:ring-ring",
-        className,
-      )}
-      {...props}
-    >
-      {options.map(({ label, value }) => (
-        <option key={value} value={value}>
-          {label}
-        </option>
-      ))}
-    </select>
-  );
-});
-
-export default Select;
